+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## IN PROGRESS
+
+* PUT on `/users?recursive=true` added, to provide a way to update agent for a specific user.
+
 ## 23.04
 
 * PUT on `/users?recursive=true` updated, to provide a way to update lines for a specific user.